refactor: separete wiki_action responsibilities
