+++ lua/neowiki/state.lua
@@ -1,10 +1,9 @@
--- lua/neowiki/state.lua
-
+--- lua/neowiki/state.lua
 ---@class neowiki.State
 ---@field public processed_wiki_paths table A list of processed wiki root path objects.
 ---@field public index_name string|nil The base name of the index file (e.g., "index").
 ---@field public markdown_extension string|nil The extension for markdown files (e.g., ".md").
-local M = {
+local State = {
   -- A list of all discovered wiki roots, sorted by path length descending
   -- to ensure the most specific path is matched first.
   -- Each item is a table: { resolved = "...", normalized = "..." }
@@ -15,4 +14,4 @@ local M = {
   markdown_extension = nil,
 }
 
-return M
+return State
